clinic/requirements.txt
Django~=3.2.9
djangorestframework~=3.12.4
djangorestframework_simplejwt==4.5.0
jwt~=1.3.1
PyJWT==2.0.0a1